+++ orb/Projects.cpp
@@ -3,56 +3,56 @@
 #include <string.h>
 
 Projects::Projects()
-{	
-	numberProjects = 0;
+{
+    numberProjects = 0;
 }
 
 /*
-	Add project to the array. 
+	Add project to the array.
 	If project already exists, replace it instead.
 */
 void Projects::add(Project project)
 {
-	if (numberProjects < MAX_NUMBER_PROJECTS)
-	{
-		int index;
+    if (numberProjects < MAX_NUMBER_PROJECTS)
+    {
+        int index;
 
-		// Already exist?
-		if ((index = find(project)) >= 0)
-		{
-			all[index] = project;
-		}
-		else
-		{
-			all[numberProjects] = project;
-			numberProjects++;
-		}
-	} 
+        // Already exist?
+        if ((index = find(project)) >= 0)
+        {
+            all[index] = project;
+        }
+        else
+        {
+            all[(unsigned int)numberProjects] = project;
+            numberProjects++;
+        }
+    }
 }
 
 int Projects::count()
 {
-	return numberProjects;
+    return numberProjects;
 }
 
 void Projects::clear()
 {
-	numberProjects = 0;
+    numberProjects = 0;
 }
 
 Project Projects::mostRecentBuild()
 {
-	int mostRecentIndex = 0;
+    int mostRecentIndex = 0;
 
-	for (int i=0; i < numberProjects; i++)
-	{
-		if (all[i].lastBuildTimeUTC > all[mostRecentIndex].lastBuildTimeUTC)
-		{
-			mostRecentIndex = i;
-		}
-	}
+    for (int i=0; i < numberProjects; i++)
+    {
+        if (all[i].lastBuildTimeUTC > all[mostRecentIndex].lastBuildTimeUTC)
+        {
+            mostRecentIndex = i;
+        }
+    }
 
-	return all[mostRecentIndex];
+    return all[mostRecentIndex];
 }
 
 /*
@@ -60,15 +60,15 @@ Project Projects::mostRecentBuild()
 */
 int Projects::building()
 {
-	for (int i=0; i < numberProjects; i++)
-	{
-		if (all[i].building)
-		{
-			return i;
-		}
-	}
+    for (int i=0; i < numberProjects; i++)
+    {
+        if (all[i].building)
+        {
+            return i;
+        }
+    }
 
-	return -1;
+    return -1;
 }
 
 /*
@@ -77,15 +77,15 @@ int Projects::building()
 int Projects::find(Project project)
 {
 
-	for (int i=0; i < numberProjects; i++)
-	{
-		if (all[i].hash == project.hash)
-		{
-			return i;
-		}
-	}
+    for (int i=0; i < numberProjects; i++)
+    {
+        if (all[i].hash == project.hash)
+        {
+            return i;
+        }
+    }
 
-	return -1;	
+    return -1;
 }
 
 /*
@@ -93,13 +93,13 @@ int Projects::find(Project project)
 */
 bool Projects::hasFailure()
 {
-	for (int i=0; i < numberProjects; i++)
-	{
-		if (!all[i].success)
-		{
-			return true;
-		}
-	}
+    for (int i=0; i < numberProjects; i++)
+    {
+        if (!all[i].success)
+        {
+            return true;
+        }
+    }
 
-	return false;
+    return false;
 }
